% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_preds}
\alias{get_preds}
\title{Get Predictions from Learner}
\usage{
get_preds(trained, task, subset = NULL)
}
\arguments{
\item{trained}{\class{WrappedModel}; A \code{mlr} learner that has been trained with \code{mlr::train} and results in a \class{WrappedModel}}

\item{task}{\class{Task}; A \code{mlr} task}

\item{subset}{<numeric vector>; A numeric vector of the rows in the \class{Task} data that should be used for training}
}
\description{
Get Predictions from Learner
}
\details{
Function to extract predictions as a list from the various individually trained learners
}
