% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnls_cvrisk}
\alias{nnls_cvrisk}
\title{nnls_cvrisk}
\usage{
nnls_cvrisk(Z, Y, algnames, verbose = T)
}
\arguments{
\item{Z}{numeric matrix; the matrix of predictions from the K algorithms}

\item{Y}{numeric; outcome}

\item{algnames}{<character vector>; names of the K algorithms that correspond to the columns in \code{Z}}

\item{verbose}{logical; Talk or quiet}
}
\description{
nnls_cvrisk
}
\details{
This is the Cross validated risk.

Note the function is essentially  lifted directly from SuperLearner::method.NNLS. I dropped observation weights and changed the return slightly
}
