% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SL_crossval_risk_pred}
\alias{SL_crossval_risk_pred}
\title{Super Learner with Cross-Validated Risk Prediction}
\usage{
SL_crossval_risk_pred(learnerlib, task, valset.list)
}
\arguments{
\item{learnlib}{list; A list of \code{Learners} from the \code{mlr} package}

\item{Task}{\class{Task}; A \class{Task} from the \code{mlr} package}

\item{valset.matrix}{list; A list of the row numbers that correspond to the validation observations for each of the V-fold processes. Note, each list item should be a vector of rownumbers}
}
\description{
Super Learner with Cross-Validated Risk Prediction
}
\details{
Please see van der Laan et al. 2007 (PMID: 17910531) & Gruber et al. 2015 (PMC4262745) for more detail
}
